+++ dubbo-common/src/test/java/org/apache/dubbo/common/threadpool/MemorySafeLinkedBlockingQueueTest.java
@@ -14,19 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.apache.dubbo.common.threadpool;
 
 import org.apache.dubbo.common.concurrent.AbortPolicy;
 import org.apache.dubbo.common.concurrent.RejectException;
 
+import java.lang.instrument.Instrumentation;
+import java.util.concurrent.LinkedBlockingQueue;
+
 import net.bytebuddy.agent.ByteBuddyAgent;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-
-import java.lang.instrument.Instrumentation;
-import java.util.concurrent.LinkedBlockingQueue;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -37,43 +36,42 @@ class MemorySafeLinkedBlockingQueueTest {
     void test() {
         ByteBuddyAgent.install();
         final Instrumentation instrumentation = ByteBuddyAgent.getInstrumentation();
-        final long objectSize = instrumentation.getObjectSize((Runnable) () -> {
-        });
-        int maxFreeMemory = (int) MemoryLimitCalculator.maxAvailable();
+        final long objectSize = instrumentation.getObjectSize((Runnable) () -> {});
+        long maxFreeMemory = (long) MemoryLimitCalculator.maxAvailable();
         MemorySafeLinkedBlockingQueue<Runnable> queue = new MemorySafeLinkedBlockingQueue<>(maxFreeMemory);
         // all memory is reserved for JVM, so it will fail here
-        assertThat(queue.offer(() -> {
-        }), is(false));
+        assertThat(queue.offer(() -> {}), is(false));
 
         // maxFreeMemory-objectSize Byte memory is reserved for the JVM, so this will succeed
         queue.setMaxFreeMemory((int) (MemoryLimitCalculator.maxAvailable() - objectSize));
-        assertThat(queue.offer(() -> {
-        }), is(true));
+        assertThat(queue.offer(() -> {}), is(true));
     }
 
     @Test
     void testCustomReject() {
-        MemorySafeLinkedBlockingQueue<Runnable> queue = new MemorySafeLinkedBlockingQueue<>(Integer.MAX_VALUE);
+        MemorySafeLinkedBlockingQueue<Runnable> queue = new MemorySafeLinkedBlockingQueue<>(Long.MAX_VALUE);
         queue.setRejector(new AbortPolicy<>());
-        assertThrows(RejectException.class, () -> queue.offer(() -> {
-        }));
+        assertThrows(RejectException.class, () -> queue.offer(() -> {}));
     }
 
     @Test
     @Disabled("This test is not stable, it may fail due to performance (C1, C2)")
     void testEfficiency() throws InterruptedException {
-        // if length is vert large(unit test may runs for a long time), so you may need to modify JVM param such as : -Xms=1024m -Xmx=2048m
-        // if you want to test efficiency of MemorySafeLinkedBlockingQueue, you may modify following param: length and times
+        // if length is vert large(unit test may runs for a long time), so you may need to modify JVM param such as :
+        // -Xms=1024m -Xmx=2048m
+        // if you want to test efficiency of MemorySafeLinkedBlockingQueue, you may modify following param: length and
+        // times
         int length = 1000, times = 1;
 
         // LinkedBlockingQueue insert Integer: 500W * 20 times
         long spent1 = spend(new LinkedBlockingQueue<>(), length, times);
 
         // MemorySafeLinkedBlockingQueue insert Integer: 500W * 20 times
-        long spent2 = spend(newMemorySafeLinkedBlockingQueue(),  length, times);
+        long spent2 = spend(newMemorySafeLinkedBlockingQueue(), length, times);
         System.gc();
 
-        System.out.println(String.format("LinkedBlockingQueue spent %s millis, MemorySafeLinkedBlockingQueue spent %s millis", spent1, spent2));
+        System.out.println(String.format(
+                "LinkedBlockingQueue spent %s millis, MemorySafeLinkedBlockingQueue spent %s millis", spent1, spent2));
         // efficiency between LinkedBlockingQueue and MemorySafeLinkedBlockingQueue is very nearly the same
         Assertions.assertTrue(spent1 - spent2 <= 1);
     }
@@ -107,7 +105,7 @@ class MemorySafeLinkedBlockingQueueTest {
     private static MemorySafeLinkedBlockingQueue<Integer> newMemorySafeLinkedBlockingQueue() {
         ByteBuddyAgent.install();
         final Instrumentation instrumentation = ByteBuddyAgent.getInstrumentation();
-        final long objectSize = instrumentation.getObjectSize((Runnable) () -> { });
+        final long objectSize = instrumentation.getObjectSize((Runnable) () -> {});
         int maxFreeMemory = (int) MemoryLimitCalculator.maxAvailable();
         MemorySafeLinkedBlockingQueue<Integer> queue = new MemorySafeLinkedBlockingQueue<>(maxFreeMemory);
         queue.setMaxFreeMemory((int) (MemoryLimitCalculator.maxAvailable() - objectSize));
